Refactoring main folder names and classes
