+++ BlockBusterApp/Form1.cs
@@ -10,7 +10,7 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 
-namespace WindowsFormsApp1
+namespace BlockBusterApp
 {
     public partial class Form1 : Form
     {
