+++ BlockBusterApp/Program.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 
-namespace WindowsFormsApp1
+namespace BlockBusterApp
 {
     static class Program
     {
